requirements.txt
streamlit>=1.30
google-genai
